// pocs/poc6-scale/src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Creature = __t.object("Creature", {
  id: __t.u64(),
  owner: __t.identity(),
  x: __t.f32(),
  y: __t.f32(),
  targetX: __t.f32(),
  targetY: __t.f32(),
  speed: __t.f32(),
  health: __t.f32(),
  chunkX: __t.i32(),
  chunkY: __t.i32(),
});
export type Creature = __Infer<typeof Creature>;

export const Player = __t.object("Player", {
  identity: __t.identity(),
  name: __t.string(),
  online: __t.bool(),
});
export type Player = __Infer<typeof Player>;

export const TickMetrics = __t.object("TickMetrics", {
  id: __t.u64(),
  tickNumber: __t.u64(),
  creatureCount: __t.u32(),
  durationMicros: __t.u64(),
  combatEvents: __t.u32(),
});
export type TickMetrics = __Infer<typeof TickMetrics>;

export const TickSchedule = __t.object("TickSchedule", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type TickSchedule = __Infer<typeof TickSchedule>;